Bloqueios de adicao e delecao. Pagina de teste recebe traducao do servidor.

+++ regex-web/WebContent/js/deleteFunctions.js
@@ -12,13 +12,16 @@
  */
 function removeNode(){
 	
-	//Recupera o elemento atualmente selecionado na arvore
+	//Recupera o ID do elemento atualmente selecionado na arvore
 	var currentSelected = getCurrentSelectedNode();
 	
 	//Se nenhum elemento estiver selecionado, nao faz nada.
 	if (currentSelected == "0") return;
 	
-	//Deleta o item selecionado da arvore, sem selecionar um novo.
-	//http://docs.dhtmlx.com/api__dhtmlxtree_deleteitem.html
-	tree.deleteItem(currentSelected, false);
+	if (canRemoveNode(currentSelected)){
+		
+		//Deleta o item selecionado da arvore, sem selecionar um novo.
+		//http://docs.dhtmlx.com/api__dhtmlxtree_deleteitem.html
+		tree.deleteItem(currentSelected, false);	
+	}
 }
